+++ mdObject/src/consts/consts.ts
@@ -1,2 +1,3 @@
 export { version } from './version';
 export { productType } from './productType';
+export { emptyImage } from './strings';