+++ mdObject/src/test/classes/Patient.spec.ts
@@ -1,5 +1,5 @@
-import { Patient, Measurements, Allergies, ReferringProvider, Phone, Address, Emr } from '../../classes/classes';
-import { MockEmrMel, mockResultEmr, MockWindow, mockResultFlowsheet } from '../mocks/mocks';
+import { Patient, Measurements, Allergies, ReferringProvider, Phone, Address } from '../../classes/classes';
+import { MockEmrMel, mockResultEmr, mockResultFlowsheet } from '../mocks/mocks';
 
 describe('Class: Patient', () => {
 
@@ -7,14 +7,13 @@ describe('Class: Patient', () => {
     let mockEmrMel = new MockEmrMel();
     let result: string = mockResultEmr;
     let resultFlowsheet: string = mockResultFlowsheet;
-    let _window = new MockWindow();
     let weight = '1';
     let height = '2';
     let name = 'name';
 
     describe('get from cache', () => {
         beforeAll(() => {
-            component = new Patient(weight, height, _window, _window.document, mockEmrMel as any);
+            component = new Patient(weight, height, mockEmrMel as any);
             component.patientId;
             component.pid;
             component.medicalRecordId;
@@ -125,7 +124,7 @@ describe('Class: Patient', () => {
 
     describe('get from mel', () => {
         beforeEach(() => {
-            component = new Patient(weight, height, _window, _window.document, mockEmrMel as any);
+            component = new Patient(weight, height, mockEmrMel as any);
             spyOn(mockEmrMel, 'melFunc')
                 .and
                 .returnValue(result);
@@ -217,12 +216,11 @@ describe('Class: Patient', () => {
         it('referringProvider', () => { expect(component.referringProvider instanceof ReferringProvider).toEqual(true); })
         it('phone', () => { expect(component.phone instanceof Phone).toEqual(true); })
         it('address', () => { expect(component.address instanceof Address).toEqual(true); })
-        it('emr', () => { expect(component.emr instanceof Emr).toEqual(true); })
     })
 
     describe('methods', () => {
         beforeEach(() => {
-            component = new Patient(weight, height, _window, _window.document, mockEmrMel as any);
+            component = new Patient(weight, height, mockEmrMel as any);
             
         });